Feature: Initial Rapidash structure with HTTP and OAuth
This initial version of Rapidash includes enough functionality to
be able to perform get requests from a RESTful API using either OAuth2
or simple HTTP.

Currently, all objects are returned as Hashie::Mash objects, however
this will be configurable in future.

JSON is the only supported format at the moment.

Resources can be defined by extending Rapidash::Base  it is required
to set a url on the resource, this could perhaps be modified to be
determined by the class name.

    class Users < Rapidash::Base
      url :users
    end

This will create an object, it needs a client to be able to call
the URL.  The code may look like:

    class Client < Rapidash::Client
      method :http
      resource :users
    end

This will allow a user to to create a client and call the following:

    client = Client.new
    client.site = "http://example.com/"
    client.users #Returns an instance of Users
    client.users! #Will make a call to "http://example.com/users
    client.users!(1) #Will make a call to "http://example.com/users/1