Feature: Support for 1.8.7, 1.9.2 and 1.9.3
The ruby 2.0.0 Hash syntax was used in the creation of an options has
in some of the tests.  This has been changed from `Hash` to `{}` to
ensure the tests work in ruby 1.9

Ruby 1.8.7 returns the methods as an array of strings instead of an
array of symbols.  Many of the tests were checking for the presence
of a particular method, so in those tests, the methods array is explicitly
converted to an array of symbols.

Ruby 1.8.7 also does not support looking up on the first character of a
string.

    "foo"[0]   # = 102
    "foo"[0,1] # = "f"

This was used to determine the response code from the HTTP request and
has been changed to support the 1.8.7 syntax.

Ruby 1.8.7 doesn't include JSON by default, so the JSON gem has been
added as a dependency and rubygems is required at the start of rapidash
and also in the tests.