Feature: Support for create, update and delete
The Base class now has methods to facilitate the creation, updating
and deletion of items.

These are called like so:

Create:

    client.posts.create!({
      :post => {
        :title => "Some post"
      }
    })

Update:

    client.posts(1).update!({
      :post => {
        :title => "Some post"
      }
    })

Delete:

    client.posts(1).delete!

Delete is fine.  I would like to include a way to specify the structure
of an object for create and post to allow either the skipping of the root
object (:post) in this case, or the need to provide a Hash at all.

Perhaps enable something like:

    client.posts.create!("Some post")