Feature: Use singular class name on a resource
When defining a resource, the singular class name is preferred to be
more in line with rails.

    class User < Rapidash::Base
    end

    class Client > Rapidash::Client
      resource :users
    end

This will set up a valid link between the client and the User class.
If the singular does not exist then the plural term can be used.  This
will raise a deprecation warning but will not break.

It will be a requirement in the future to let a user override the class
when defining a resource:

    class Client > Rapidash::Client
      resource :admin_users, :class_name => Users
    end