Feature: Allow a nil response to be parsed
It is a fairly common rails paradigm to respond with nil on a create
or update instead of the object that was modified.  Something like:

    def create
      some_model.save
      respond_with(nil)
    end

This will generate the following json: `null`

This is technically not valid JSON, as it is a single level object
however in Rapidash it will return nil instead of raising an error.